pyproject.toml
[project]
name = "dashcorn"
version = "0.1.0"
description = "Flower-like dashboard for FastAPI/Uvicorn"
authors = [
    { name = "dashcorn" }
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic>2.0",
    "fastapi[standard]<1.0.0,>=0.114.2",
    "httpx<1.0.0,>=0.25.1",
    "uvicorn==0.32.0",
    "pyzmq>=27.0.0",
]

[dependency-groups]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "pytest-cov<4.0.0,>=3.0.0",
    "pydash>=8.0.5",
    "mypy<2.0.0,>=1.8.0",
]
lint = [
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
]
